remove special node
