+++ pgdrive/component/vehicle_module/lidar.py
@@ -1,4 +1,5 @@
 from typing import Set
+from pgdrive.utils.utils import get_object_from_node
 
 from panda3d.core import BitMask32, NodePath
 
@@ -26,7 +27,7 @@ class Lidar(DistanceDetector):
             if ret.getNode().hasPythonTag(BodyName.Traffic_vehicle):
                 vehicles.add(ret.getNode().getPythonTag(BodyName.Traffic_vehicle).kinematic_model)
             elif ret.getNode().hasPythonTag(BodyName.Base_vehicle):
-                vehicles.add(ret.getNode().getPythonTag(BodyName.Base_vehicle))
+                vehicles.add(get_object_from_node(ret.getNode()))
         return vehicles
 
     # def _get_surrounding_objects(self) -> Set[Object]:
